03_demo-bot/requirements.txt
pyTelegramBotAPI==4.16.1
telebot==0.0.5
SQLAlchemy==2.0.28
psycopg2==2.9.9
openpyxl==3.1.2
pandas==2.2.1
prettytables
